+++ trunk/Ventcore/src/ventcore/client/InviteReceivedDialog.java
@@ -1,5 +1,7 @@
 package ventcore.client;
 
+import wired.event.User;
+
 import com.allen_sauer.gwt.voices.client.Sound;
 import com.google.gwt.event.dom.client.ClickEvent;
 import com.google.gwt.event.dom.client.ClickHandler;
@@ -12,7 +14,7 @@ import com.google.gwt.user.client.ui.VerticalPanel;
 public class InviteReceivedDialog extends DialogBox {
 	private static final Sound inviteSound = Ventcore.createSound("invite");
 
-	public InviteReceivedDialog(User user, final int chatId) {
+	public InviteReceivedDialog(User user, final long chatId) {
 		setText("Private Chat Invitation");
 		setAnimationEnabled(true);
 		VerticalPanel p = new VerticalPanel();
